requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.0
ConfigArgParse==1.7
constantly==23.10.4
cryptography==44.0.3
daphne==4.1.2
Django==5.1.7
django-environ==0.12.0
Flask==3.1.1
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.37.3
locust-cloud==1.21.5
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==25.0
paho-mqtt==2.1.0
platformdirs==4.3.8
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
python-engineio==4.12.1
python-socketio==5.13.0
pytz==2025.2
pyzmq==26.4.0
redis==6.0.0
requests==2.32.3
service-identity==24.2.0
simple-websocket==1.1.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.9.0
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
